tests store setup
